+++ README.md
@@ -7,7 +7,6 @@ Software for high density electrophysiology.
 
  - [onlineDetection](onlineDetection): Online-capable spike detection, done independently on each recording channel.
  - [interpolatingDetection](interpolatingDetection): Spike detection with spatial interpolation. Returns cut-outs for detected events from multiple channels, which allows performing spike localisation.
- - [InterpDetectionDev](InterpDetectionDev): Development version of the interpolation and spike localisation code.
  - [postProcessing](postProcessing): Programs for removing duplicate events in spikes detected with the online method, and localise spikes detected with the interpolation method.
  - [clustering](clustering): Perform spike sorting by location and PCA on interpolated data.
  - [visualisationtool](visualisationtool): A basic GUI tool for visualising and annotating sorted spikes.
@@ -25,17 +24,21 @@ Main features are a robust noise estimate based on signal percentiles rather tha
 
 This process is fully automated, as manual inspection gets very time consuming for large-scale recordings. The following steps lead all the way from raw data to sorted units:
 
-1. Spike detection and spatial signal interpolation - currently runs in 0.1 * real time for 4000 channels, scales linearly with recording duration). A stable version is in the sub-project [interpolatingDetection](interpolatingDetection), more efficient is the code in [InterpDetectionDev](InterpDetectionDev).
+1. Spike detection and spatial signal interpolation - currently runs in 0.1 * real time for 4000 channels, scales linearly with recording duration). Code is in the sub-project [interpolatingDetection](interpolatingDetection).
 
 2. Spatial event localisation, now has about real-time performance (scales linearly with event number).
-The stable version is in the sub-project  [postProcessing](postProcessing), more efficient is the code in [InterpDetectionDev](InterpDetectionDev).
+The relevant code is in the sub-project  [postProcessing](postProcessing).
 
 3. Clustering, code in [clustering](clustering). This step is parallelised and extremely efficient, typical experiments with millions of spikes are clustered in minutes. The code also includes functions for noise removal and quality control.
 
 These methods have for far only been tested with data recorded with the [3Brain Biocam](http://www.3brain.com/biocam-system), we would be keen to other data sets. Steps 1 and 2 would have to be adjusted for other systems, step 3 is already rather generic.
 
-## Contributors
+# Contributors
 - [Matthias Hennig](http://homepages.inf.ed.ac.uk/mhennig/index.html): Spike sorting
 - [Oliver Muthmann](mailto:ollimuh@googlemail.com): Spike detection and localisation
 - [Martino Sorbaro](http://martinosorb.github.io): Spike sorting
 - [Cesar Juarez Ramirez](mailto:cesaripn2@gmail.com): Visualisation toolkit
+
+# References
+
+J.-O. Muthmann, H. Amin, E. Sernagor, A. Maccione, D. Panas, L. Berdondini, U.S. Bhalla, M.H. Hennig MH (2015). [Spike detection for large neural populations using high density multielectrode arrays](http://journal.frontiersin.org/article/10.3389/fninf.2015.00028/abstract). Front. Neuroinform. 9:28. doi: 10.3389/fninf.2015.00028.
