+++ clustering/README.md
@@ -1,7 +1,7 @@
 Spike sorting
 =============
 
-This code performs fully automated spike sorting on events detected and localised with the interpolating spike detection tools in  the sub-project [postProcessing](../postProcessing) or the development version under [InterpDetectionDev](../InterpDetectionDev).
+This code performs fully automated spike sorting on events detected and localised with the interpolating spike detection tools in  the sub-project [postProcessing](../postProcessing).
 
 
 The full workflow is illustrated in the jupyter notebook [Spike sorting workflow demonstration.ipynb](Spike sorting workflow demonstration.ipynb).
