+++ interpolatingDetection/stable/Main_reverse.cs
