+++ interpolatingDetection/stable/README.md
@@ -1,0 +1,10 @@
+# Interpolating spike detection
+
+This directory contains older, thoroughly tested code for reference.
+
+Developed and written by [Oliver Muthmann](ollimuh@googlemail.com).
+
+Reference: J.-O. Muthmann, H. Amin, E. Sernagor, A. Maccione, D. Panas, L. Berdondini, U.S. Bhalla, M.H. Hennig MH (2015). [Spike detection for large neural populations using high density multielectrode arrays](http://journal.frontiersin.org/article/10.3389/fninf.2015.00028/abstract). Front. Neuroinform. 9:28. doi: 10.3389/fninf.2015.00028.
+
+
+
