+++ interpolatingDetection/stable/SpkDinterp.exe
