+++ interpolatingDetection/stable/SpkDreverse.exe
